+++ generators/app/inputs/appDescription.js
@@ -10,6 +10,8 @@
 (function () {
 	'use strict';
 
+	const logs = require('../../common/logs.js');
+
 	const prompts = [
 		{
 			type   : 'input',
@@ -24,13 +26,14 @@
 	];
 
 	module.exports = $that => {
-		$that.logHintHeader();
-		$that.logHints(hints);
+		logs.hintHeader($that);
+		logs.hints($that, hints);
 
 		prompts[0].default += ' ' + $that.appName;
 
 		return $that.prompt(prompts).then($response => {
 			$that.appDescription = $response.appDescription;
+			$that.config.set('appDescription', $that.appDescription);
 			$that.log();
 		});
 	};