+++ generators/app/inputs/author.js
@@ -10,6 +10,8 @@
 (function () {
 	'use strict';
 
+	const logs = require('../../common/logs.js');
+
 	const prompts = [
 		{
 			type    : 'input',
@@ -41,8 +43,8 @@
 
 	module.exports = class Author {
 		static set($that) {
-			$that.logHintHeader();
-			$that.logHints(hints);
+			logs.hintHeader($that);
+			logs.hintsMultiline($that, hints);
 
 			return $that.prompt(prompts).then($response => {
 				this.onSuccess($that, $response);
@@ -52,6 +54,9 @@
 		static onSuccess($that, $response) {
 			$that.authorFirstName = $response.authorFirstName;
 			$that.authorLastName  = $response.authorLastName;
+			$that.config.set('authorFirstName', $that.authorFirstName);
+			$that.config.set('authorLastName', $that.authorLastName);
+			$that.log();
 		}
 	};