+++ generators/app/inputs/authorEmail.js
@@ -10,7 +10,8 @@
 (function () {
 	'use strict';
 
-	const _ = require('lodash');
+	const _    = require('lodash');
+	const logs = require('../../common/logs.js');
 
 	const prompts = [
 		{
@@ -25,8 +26,8 @@
 	];
 
 	module.exports = $that => {
-		$that.logHintHeader();
-		$that.logHints(hints);
+		logs.hintHeader($that);
+		logs.hints($that, hints);
 
 		prompts[0].default = _.lowerCase($that.authorFirstName) + '.' + _.lowerCase($that.authorLastName) + '@cozen.com';
 
@@ -34,6 +35,9 @@
 			$that.authorEmail = $response.authorEmail;
 			$that.authorShort = $that.authorFirstName + ' ' + $that.authorLastName;
 			$that.authorLong  = $that.authorFirstName + ' ' + $that.authorLastName + ' ' + $that.authorEmail;
+			$that.config.set('authorEmail', $that.authorEmail);
+			$that.config.set('authorShort', $that.authorShort);
+			$that.config.set('authorLong', $that.authorLong);
 			$that.log();
 		});
 	};