+++ generators/app/inputs/lang.js
@@ -15,8 +15,9 @@
 		level  : 3,
 		enabled: true
 	});
-	const colors        = require('../colors.js');
+	const colors        = require('../../common/colors.js');
 	const _             = require('lodash');
+	const logs          = require('../../common/logs.js');
 
 	const prompts = [
 		{
@@ -33,15 +34,17 @@
 	];
 
 	module.exports = $that => {
-		$that.logHintHeader();
-		$that.logHints(hints);
+		logs.hintHeader($that);
+		logs.hints($that, hints);
 
 		$that.log('Current lang list :');
 		$that.log(chalk.hex(colors.get('cyan'))('fr'));
 		$that.log(chalk.hex(colors.get('cyan'))('en'));
+		$that.log();
 
 		return $that.prompt(prompts).then($response => {
 			$that.lang = _.lowerCase($response.lang);
+			$that.config.set('lang', $that.lang);
 			$that.log();
 		});
 	};