+++ generators/view/templates/constant.js
@@ -1,0 +1,11 @@
+/**
+ <%- include('../../common/partial/header.txt') %>
+ */
+(function (angular) {
+	'use strict';
+
+	angular
+		.module('<%= appNameCamel %>')
+		.constant('<%= constantName %>', {});
+
+})(window.angular);